MH-8100 updated mediaelement functions to return long instead of int

+++ modules/matterhorn-videoeditor-impl/src/main/java/org/opencastproject/videoeditor/impl/VideoEditorService.java
@@ -76,8 +76,8 @@ public class VideoEditorService implements VideoEditor, ManagedService {
 
     for (ParallelElement pe : smil.getBody().getSequence().getElements()) {
       for (MediaElement me : pe.getElements()) {
-        int begin = me.getClipBeginMS();
-        int end = me.getClipEndMS();
+        long begin = me.getClipBeginMS();
+        long end = me.getClipEndMS();
         String srcFile = me.getSrc();
 
         try {
